Removed GOPATH and PATH assignment for go directories
I have added a much better solution for GOPATH and PATH in the file gopath-dir-locals.el.  Just symlink that file to your GOPATH/.dir-locals.el and then, whenever you open a file in any sub-directory of that directory, your GOPATH and exec-path settings will be updated correctly.

I have similar logic that I added to my .zshrc so that when I change directories, it will recursively look for a .gopath file to symbolize that the directory is a GOPATH, and set up the environment correspondingly.